requirements.txt
playwright>=1.45.0
pdfplumber>=0.11.0
reportlab>=4.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
